scripts/mcp_servers/pdf_reader_mcp/requirements.txt
pymupdf>=1.24.0
google-generativeai>=0.8.0
Pillow>=10.0.0
aiofiles>=24.0.0
python-dotenv>=1.0.0

